+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   in files
 * Fix bug in `ZeroUnit` linter where a string with a substring that looked
   like a zero followed by units would incorrectly report a lint
+* Fix bug in `TrailingSemicolonAfterPropertyValue` where a lint would be
+  incorrectly reported for properties split up over multiple lines
 
 ## 0.21.0